pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgenii_Lobachev",email = "lobachev_e_96@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
mypy = "^1.17.1"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
python-dateutil = "^2.9.0.post0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.git
      \.idea
      \.venv
  )/
)
'''